+++ scheduling/test-suite/src/lib.rs
@@ -1,9 +1,9 @@
 mod access;
 mod scheduler_ext;
-mod transaction;
+mod tx;
 mod vm;
 
 pub use access::Access;
 pub use scheduler_ext::SchedulerExt;
-pub use transaction::Tx;
+pub use tx::Tx;
 pub use vm::VM;
